added nested pojo
