+++ app/src/main/java/com/androidvoyage/matrimonk/database/MatchItem.kt
@@ -5,6 +5,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Embedded
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import com.androidvoyage.matrimonk.database.pojo.*
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
@@ -12,7 +13,7 @@ import kotlinx.android.parcel.Parcelize
 data class MatchItem(
 
         @PrimaryKey(autoGenerate = true)
-        var matchId: Long? = 0L,
+        var matchId: Long = 0L,
 
         @ColumnInfo(name = "gender")
         var gender: String? = "Male",
@@ -29,11 +30,8 @@ data class MatchItem(
         @ColumnInfo(name = "nat")
         var nat: String? = "",
 
-        /*@Embedded
-        var name: Name? = Name(),
-
         @Embedded
-        var location: Location? = Location(),
+        var name: Name? = Name(),
 
         @Embedded
         var login: Login? = Login(),
@@ -41,117 +39,18 @@ data class MatchItem(
         @Embedded
         var dob: Dob? = Dob(),
 
-        @Embedded
-        var registered: Registered? = Registered(),
-
-        @Embedded
+        @Embedded(prefix = "id_")
         var id: Id? = Id(),
 
         @Embedded
-        var picture: Picture? = Picture(),*/
+        var picture: Picture? = Picture(),
 
-        ) : Parcelable
+        @Embedded(prefix = "reg_")
+        var registered: Registered? = Registered(),
 
-@Parcelize
-data class Name(
-        @ColumnInfo(name = "title")
-        var title: String? = "",
-        @ColumnInfo(name = "first")
-        var first: String? = "",
-        @ColumnInfo(name = "last")
-        var last: String? = ""
-) : Parcelable
+        @Embedded(prefix = "loc_")
+        var location: Location? = Location()
 
-@Parcelize
-class Street(
-        @ColumnInfo(name = "number")
-        var number: Long = 0,
-        @ColumnInfo(name = "street_name")
-        var name: String? = ""
-) : Parcelable
 
-@Parcelize
-class Coordinates(
-        @ColumnInfo(name = "latitude")
-        var latitude: String? = "",
-        @ColumnInfo(name = "longitude")
-        var longitude: String? = ""
-) : Parcelable
 
-@Parcelize
-class Timezone(
-        @ColumnInfo(name = "offset")
-        var offset: String? = "",
-        @ColumnInfo(name = "description")
-        var description: String? = ""
-) : Parcelable
-
-@Parcelize
-class Location(
-        @ColumnInfo(name = "street")
-        var street: Street? = Street(),
-        @ColumnInfo(name = "city")
-        var city: String? = "",
-        @ColumnInfo(name = "state")
-        var state: String? = "Maharashtra",
-        @ColumnInfo(name = "country")
-        var country: String? = "INDIA",
-        @ColumnInfo(name = "postcode")
-        var postcode: String? = "",
-        @ColumnInfo(name = "coordinates")
-        var coordinates: Coordinates? = Coordinates(),
-        @ColumnInfo(name = "timezone")
-        var timezone: Timezone? = Timezone()
-) : Parcelable
-
-@Parcelize
-class Login(
-        @ColumnInfo(name = "uuid")
-        var uuid: String? = "",
-        @ColumnInfo(name = "username")
-        var username: String? = "",
-        @ColumnInfo(name = "password")
-        var password: String? = "",
-        @ColumnInfo(name = "salt")
-        var salt: String? = "",
-        @ColumnInfo(name = "md5")
-        var md5: String? = "",
-        @ColumnInfo(name = "sha1")
-        var sha1: String? = "",
-        @ColumnInfo(name = "sha256")
-        var sha256: String? = ""
-) : Parcelable
-
-@Parcelize
-class Dob(
-        @ColumnInfo(name = "date")
-        var date: String? = "",
-        @ColumnInfo(name = "age")
-        var age: Int? = 0
-) : Parcelable
-
-@Parcelize
-class Registered(
-        @ColumnInfo(name = "join_date")
-        var date: String? = "",
-        @ColumnInfo(name = "join_age")
-        var age: String? = ""
-) : Parcelable
-
-@Parcelize
-class Id(
-        @ColumnInfo(name = "id_name")
-        var name: String? = "",
-        @ColumnInfo(name = "id_value")
-        var value: String? = ""
-) : Parcelable
-
-@Parcelize
-class Picture(
-        @ColumnInfo(name = "large")
-        var large: String? = "",
-        @ColumnInfo(name = "medium")
-        var medium: String? = "",
-        @ColumnInfo(name = "thumbnail")
-        var thumbnail: String? = ""
 ) : Parcelable
