+++ app/src/main/java/com/androidvoyage/matrimonk/database/pojo/Dob.kt
@@ -1,0 +1,10 @@
+package com.androidvoyage.matrimonk.database.pojo
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class Dob(
+    var date: String? = "",
+    var age: Int? = 0
+) : Parcelable
