+++ app/src/main/java/com/androidvoyage/matrimonk/database/pojo/Name.kt
@@ -1,0 +1,13 @@
+package com.androidvoyage.matrimonk.database.pojo
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+
+
+@Parcelize
+data class Name(
+    var title: String? = "",
+    var first: String? = "",
+    var last: String? = ""
+) : Parcelable
