+++ app/src/main/java/com/androidvoyage/matrimonk/database/pojo/Registered.kt
@@ -1,0 +1,10 @@
+package com.androidvoyage.matrimonk.database.pojo
+
+import android.os.Parcelable
+import kotlinx.android.parcel.Parcelize
+
+@Parcelize
+class Registered(
+    var date: String? = "",
+    var age: String? = ""
+) : Parcelable
