+++ app/src/main/res/layout/item_matches.xml
@@ -21,6 +21,7 @@
         android:layout_marginStart="16dp"
         android:layout_marginTop="8dp"
         android:layout_marginEnd="16dp"
+        android:onClick="@{()->clickListener.onAccept(match)}"
         android:layout_marginBottom="8dp"
         app:cardBackgroundColor="@color/colorWhite"
         app:cardCornerRadius="16dp">
